* Fix bugs

+++ ExpenseManagementTableViewController.m
@@ -32,7 +32,7 @@ GroupTabBarController *groupTabBarController;
     [expensesQuery findObjectsInBackgroundWithBlock:^(NSArray *objects, NSError *error) {
         if (!error) {
             // The find succeeded.
-            NSLog(@"Successfully retrieved %d Expenses.", objects.count);
+            NSLog(@"Successfully retrieved %lu Expenses.", (unsigned long)objects.count);
             
             // Do something with the found objects
             groupTabBarController.expenses = [objects mutableCopy];
